Fixed varNOpp
Fixed is a strong word, the game is playing again, but I'm not sure that the script is actually printing. I think it's trying to, but I haven't seen the console... so I think it works. Anyway, the problem is that in C# you often use the line "console.WriteLine()", but for unity you replace that with "print()". so it's figured out.