Fixed Errors and added else if and else lines
So, there were some errors with the syntax script that weren't stopping the game, but I fixed those. So, now the console is printing the 3 sums from the varNOpps and the string from the Syntax script. I tried to add an else if so that if the block was about 2.5F off the ground it would print "AAAAAAHHHHH" instead of "I'mma bout to hit the ground" and an else so that if it was within 2.5 feet or about 5 feet from the ground it would say "I'm fine... yeah... I'm... fine...." but the else if and else didn't take over. Also, I haven't figured out how to print the string and char together, I want to print to console nameX + charX to do a "last name first initial (or vice versa) thing, and that's not working and googling isn't doing much.