Added ConSwitch script
I added ConSwtich script, to practice conditionals and switches. I have one conditional in there, pretty simple, just and "if," and an "else,". I have a more complicated swtich which has cases 1-5 and default. The numbers are set as static int, so i'd like to go in a make a condition for the number to be randomly selected and then run the conditional and the switch so that the other print's I wrote would get some use eventually.