added Loop script
so, added a script for loop/array. it was a little more "copy/paste" than I like. I did find a set of code to do a give a random element for an array, looked simple enough (when copy/paste), but I decided to get the assignment done and I'll implement it when I can.